+++ model/mmwave-phy.cc
@@ -290,12 +290,11 @@ void
 MmWavePhy::SendRachPreamble (uint32_t PreambleId, uint32_t Rnti)
 {
   NS_LOG_FUNCTION (this);
-  // This function is called by the SAP, SO it has to stay at the L1L2CtrlDelay rule
   m_raPreambleId = PreambleId;
   Ptr<MmWaveRachPreambleMessage> msg = Create<MmWaveRachPreambleMessage> ();
   msg->SetSourceBwp (GetBwpId ());
   msg->SetRapId (PreambleId);
-  EnqueueCtrlMessage (msg); // Enqueue at the end
+  EnqueueCtrlMsgNow (msg);
 }
 
 void